+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 11/08/2025
+
+### Added
+
+- Manage Access Key and Manage PromptTab
+- Fix route endpoint on /projects for fetching via API key
+- Removed Icon input on add skill modal for unified flow.
+- Added system_prompt table on database
+
 ## [0.1.2] - 11/05/2025
 
 ### Added